refactor(test_framework): share the no-conditional-in-test traversal
Move the traversal and its schema into
internal/utils/test_framework/rules/no_conditional_in_test and reduce
jest/no-conditional-in-test and rstest/no-conditional-in-test to
adapters that only say which CallExpression registers a test case.

The jest rule gains the two behaviors the rstest port had: registration
depth is counted rather than flagged, so an inner registration exiting
no longer clears the outer test's scope, and the reported range of an
optional chain covers a trailing `!`.
