--- internal/plugins/rstest/rules/no_conditional_in_test/no_conditional_in_test.schema.json
@@ -1,16 +1,0 @@
-{
-  "type": "array",
-  "items": [
-    {
-      "type": "object",
-      "properties": {
-        "allowOptionalChaining": {
-          "type": "boolean"
-        }
-      },
-      "additionalProperties": false
-    }
-  ],
-  "minItems": 0,
-  "maxItems": 1
-}
