+++ internal/utils/test_framework/rules/no_conditional_in_test/no_conditional_in_test.schema.json
